+++ database/migrations/2022_10_04_072222_create_potensi_pesantrens_table.php
@@ -16,8 +16,8 @@ class CreatePotensiPesantrensTable extends Migration
         Schema::create('potensi_pesantren', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->timestamps();
-            $table->integer('pesantrenid');
-            $table->integer('potensiid');
+            $table->unsignedBigInteger('pesantrenid');
+            $table->unsignedBigInteger('potensiid');
         });
     }
